[+] AddPipelineSteps
