--- src/components/containers/AddPipelineForm.scss
@@ -1,3 +1,0 @@
-.ant-form-item {
-  margin-bottom: 0px;
-}
